requirements/prod.txt
# Everything needed in production

# Flask
Flask==1.0.2
MarkupSafe==1.0
Werkzeug==0.14.1
Jinja2==2.10
itsdangerous==0.24
click>=5.0

# Database
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.5
SQLAlchemy==1.2.8
mysqlclient

# Migrations
Flask-Migrate==2.2.0

# Forms
Flask-WTF==0.14.2
WTForms==2.2.1

# Deployment
gunicorn>=19.1.1

# Webpack
flask-webpack==0.1.0

# Auth
Flask-Login==0.4.1
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar==0.10.1

# Extras
pandas
numpy
sklearn
progressbar2
faker
scipy
